backend/requirements.txt
pandas
python-dotenv
requests
sqlalchemy
flask
flask_sqlalchemy